CIRCSTORE-399: Search index fields migration script skeleton (#394)
* moving internal classes out

* moving batch counting and migration verification

* moving migration and error handling methods

* most code moved into abstract class

* adding skeleton class
